Converted DrawableMatcher to kotlin and moved to the androidtest module
