+++ collect_app/src/androidTest/java/org/odk/collect/android/support/pages/QRCodePage.java
@@ -14,7 +14,7 @@ import androidx.test.espresso.Espresso;
 
 import org.odk.collect.android.support.ActivityHelpers;
 import org.odk.collect.android.support.WaitFor;
-import org.odk.collect.android.support.matchers.DrawableMatcher;
+import org.odk.collect.androidtest.DrawableMatcher;
 
 public class QRCodePage extends Page<QRCodePage> {
     @Override
